% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_itemsPar.R
\name{best_itemsPar}
\alias{best_itemsPar}
\title{Fits all possible combinations of k items from a pool of n items to a single-factor model using parallel processing.}
\usage{
best_itemsPar(x, k)
}
\arguments{
\item{x}{A data frame containing the items. The name of the data frame must be x and it must be loaded in the global environment before the function is run.}

\item{k}{The number of items to be selected.}
}
\description{
The function can be useful in selecting items from a larger pool of items. The function allows for comparisons of all possible combinations of k items from the pool with respect to fit with a single factor model and reliability. The function works best if the number of possible unique combinations of items is less than 5000. It becomes extremely slow if there are more than 5000 possible combinations. This function employs parallel processing to speed up the process, but it remains slow.
}
\examples{
library(lordif)
data(Anxiety)
## Say we want to select the best four items from the first 15 Anxiety items
## The choose() function shows how many possible unique combinations there are
choose(ncol(Anxiety[4:15], 4))

## Store the data frame in the global environment as x
x <- Anxiety[4:15]
out <- best_itemsPar(x, 6) 
out
}
