% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_wlsmv.R
\name{cfa_wlsmv}
\alias{cfa_wlsmv}
\title{Confirmatory factor analysis of ordinal items with WLSMV estimation}
\usage{
cfa_wlsmv(x)
}
\arguments{
\item{x}{A data frame containing the items to be analysed}
}
\value{
The output is a list containing the results of the analysis.
}
\description{
Fits a one-factor CFA model with uncorrelated errors to the polychoric correlations of the items with WLSMV estimation. The function is a wrapper for the cfa() function of the lavaan package.
}
\examples{

library(psych)
data(bfi)

cfa_wlsmv(x)

}
