% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.ancova.R
\name{dif.ancova}
\alias{dif.ancova}
\title{Performs a differential item functioning analysis of polytomous items using ANCOVA}
\usage{
dif.ancova(data, Trait, Group)
}
\arguments{
\item{data}{A data frame containing the items. All the items should be scored in the same direction.}

\item{Trait}{A vector containing measures of the trait. This may be the total scores across all the items, the mean item scores, factor scores, or IRT person measures.}

\item{Group}{A factor indicating group membership}
}
\description{
Differential item functioning analysis of polytomous items via ANCOVA
}
\examples{
out <- dif.ancova(work_stress[1:9], rowSums(workStress[1:9]), work_stress$gender)
out
}
