% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-itemselectr-spare.Rmd.
\name{dif.Rasch}
\alias{dif.Rasch}
\title{Detection of statistically significant differential item functioning via ANCOVA of standardized Rasch residuals}
\usage{
dif.Rasch(std.residuals, Theta, Group)
}
\arguments{
\item{std.residuals}{A matrix or data frame containing the standardized residuals of a Rasch analysis}

\item{Theta}{A vector containing thetas (person measures)}

\item{Group}{A vector (factor) containing group membership}
}
\value{
DIF results
}
\description{
Detection of statistically significant differential item functioning via ANCOVA of standardized Rasch residuals
}