% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influential.cases.R
\name{influential.cases}
\alias{influential.cases}
\title{Identify influential cases and examine their impact}
\usage{
influential.cases(x, ncases = 1)
}
\arguments{
\item{x}{Data frame containing items}

\item{ncases}{The number of cases to treat as influential. This value can be set to 1 and then adjusted in subsequent analyses.}
}
\value{
A list containing (a) the factor loadings, (b) fit statistics of a single factor solution, and (c) the Cronbach alpha coefficients of the full and filtered data sets.
}
\description{
Calculates generalized Cook's Distance statistics. Removes a specified number of cases with largest gCD values. Compares factor loadings, fit statistics and reliabilities of full data set with the filtered data set.
}
\examples{
library(lordif)
data(Anxiety)
influential.cases(Anxiety[1:250, 4:32], 1)
}
