% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-itemselectr-spare.Rmd.
\name{info.select}
\alias{info.select}
\title{Item selection based on McDonald's item information statistic}
\usage{
info.select(x, nitems)
}
\arguments{
\item{x}{A data frame containing items}

\item{nitems}{The number of items to retain in the scale}
}
\value{
The function sequentially removes the item with the lowest information statistic from a pool of items. The output is a list containing (a) a data frame that shows at each step of the item selection process the number of items that had been removed, the name of the item that is removed, and the reliability of the scale; (b) the names of the items that were retained; (c) coefficient omega of the final scale, and (d) coefficient alpha of the final scale.
}
\description{
An item selection procedure based on item information statistics.
}
\examples{
library(lordif)
data(Anxiety)
info.select(Anxiety[4:32], 5)
}