% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.information.R
\name{item.information}
\alias{item.information}
\title{McDonald's item information statistic}
\usage{
item.information(x)
}
\arguments{
\item{x}{A data frame containing items}
}
\value{
A matrix containing for each item its (a) unstandardized factor loading, (b) square of the loading, (c) unique variance, and (d) item information.
}
\description{
Calculates item information statistics (McDonald, 1999).
}
\examples{
info <- item.information(work_stress[1:9])
}
