% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-itemselectr-spare.Rmd.
\name{item.reliability.index}
\alias{item.reliability.index}
\title{Gulliksen's item reliability index extended to polytomous items}
\usage{
item.reliability.index(x)
}
\arguments{
\item{x}{A data frame that contains the items to be analysed}
}
\value{
The item reliability index of each item
}
\description{
Takes the item-total correlation and multiplies it with the item standard deviation
}
\examples{
#Calculate item reliability indexes for the items of the Neuroticism Scale of the bfi data set
library(psychTools)
data(bfi)
item.reliability.index(bfi[16:20])
}