% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.selection.R
\name{item.selection}
\alias{item.selection}
\title{Selecting items based on Gulliksen's item reliability index}
\usage{
item.selection(x, nitems)
}
\arguments{
\item{x}{A data frame that contains the items to be analysed}

\item{nitems}{The number of items to retain. This value should never be less than three.}
}
\value{
At each step the number of items that have been removed, the name of the last item that has been removed, omega, the average item score, and the standard deviation of the test.
}
\description{
Iteratively removes item with the lowest item reliability index from a pool of items. At each step this yields a test with the largest possible variance.
}
\examples{
#Select the combination of k items that will maximise the test variance
item.selection(work_stress[1:9], 5)
}
