% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemselectr.r
\docType{package}
\name{itemselectr-package}
\alias{itemselectr}
\alias{itemselectr-package}
\title{Aids selection of single items from multi-item scales based on their psychometric properties. E.g., for use in EMA and SCED studies that require ultra brief measures.}
\description{
Fit CFA model and return factor loading and errors. Also de-attenuate item-sum score correlations by the scale internal consistency.
}
\author{
\strong{Maintainer}: Ian Hussey \email{ianhusseysipod@gmail.com} (\href{https://orcid.org/0000-0001-8906-7559}{ORCID})

}
