% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large.residuals.R
\name{large.residuals}
\alias{large.residuals}
\title{Large factor analysis residuals}
\usage{
large.residuals(x, upper = 0.08, lower = -0.08)
}
\arguments{
\item{x}{A square matrix of residuals of a factor analysis}

\item{upper}{Upper limit for acceptable residuals (defaults to 0.08)}

\item{lower}{Lower limit for acceptable residuals (defaults to -0.)}
}
\value{
Statistics relating to the residuals
}
\description{
Identifying large residuals in a matrix of factor analysis residuals
}
\examples{
library(psychTools)
data(bfi)
myresiduals <- psych::fa(bfi[1:10], nfactors = 2)$resid
myoutput <- large.residuals(x = myresiduals)
myoutput
}
