% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings_and_errors.r
\name{loadings_and_errors}
\alias{loadings_and_errors}
\title{loadings_and_errors}
\usage{
loadings_and_errors(data)
}
\arguments{
\item{data}{Item level data for a given scale. One item per column, one row per participant. No other columns other than the scale items.}
}
\description{
Fit a CFA with all parameters free. Return a list with the lavaan fit object and a table of factor loadings and intercepts along with their 95% CI
}
\examples{
loadings_and_errors(data)
}
