% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.norms.R
\name{make.norms}
\alias{make.norms}
\title{Create norms for psychological or educational tests}
\usage{
make.norms(xsum)
}
\arguments{
\item{xsum}{A vector or raw total scores for a test (no missing data)}
}
\description{
Creates percentiles, McCall T-scores (M = 50, SD = 10) and stanines (M = 5, SD = 2)
}
\examples{
# Create norms for the work_stress scale
Stotal <- rowSums(work_stress[1:9])
make.norms(Stotal)

}
