% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_item_alpha.r
\name{max_item_alpha}
\alias{max_item_alpha}
\title{max_item_alpha}
\usage{
max_item_alpha(data)
}
\arguments{
\item{data}{Item level data for a given scale. One item per column, one row per participant. No other columns other than the scale items.}
}
\description{
Calculate the max Cronbach's alpha for each item given the scale's internal consistency and each item's correlation with the sum score.
}
\examples{
max_item_alpha(data)
}
