% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_with_cis.r
\name{r_with_cis}
\alias{r_with_cis}
\title{r_with_cis}
\usage{
r_with_cis(data)
}
\arguments{
\item{data}{data frame with two columns, scale sum_scores and scores from a single item.}
}
\description{
Bootstrapped correlations between a single item and scale sum scores.
}
\examples{
r_with_cis(data)
}
