% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.DIF.R
\name{Rasch.DIF}
\alias{Rasch.DIF}
\title{Performs a differential item functioning analysis of polytomous items using ANCOVA}
\usage{
Rasch.DIF(std.res, group, thetas)
}
\arguments{
\item{std.res}{A data frame containing the standardized residuals of a Rasch analysis}

\item{group}{A factor indicating group membership}

\item{thetas}{A vector containing measures of the trait. This may be the total scores across all the items, the mean item scores, factor scores, or IRT person measures.}
}
\description{
Differential item functioning analysis of polytomous items via ANCOVA
}
\examples{
# Lowest response category must be scored as 0
work_stress0 <- work_stress[1:9] - 1
d <- Simple.Rasch(work_stress0, "PCM2")

# Store the standardized residuals as a data frame
d$std.residuals <- data.frame(d$std.residuals)

# Perform the dif analysis
out <- Rasch.DIF(d$std.residuals, work_stress$gender, d$thetas[,1])
out
}
