% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaschItemAnalysis.R
\name{RaschItemAnalysis}
\alias{RaschItemAnalysis}
\title{Rasch analysis of polytomous items with the partial credit model}
\usage{
RaschItemAnalysis(x)
}
\arguments{
\item{x}{A data frame containing items where the lowest response category is scored as 0}
}
\value{
The output is a list containing the results of the analysis.
}
\description{
Basic Rasch analysis of polytomous items using the partial credit model. This function is a wrapper for the PCM() function of the eRm package. The purpose of the function is to simplify the Rasch analysis process and to produce the output that users most typically want.
}
\examples{
# Lowest response category must be scored as 0
work_stress0 <- work_stress[1:9] - 1

## Fit the partial credit model
out <- RaschItemAnalysis(work_stress0)
ls(out)
out$Infit.outfit
}
