% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.code.R
\name{reverse.code}
\alias{reverse.code}
\title{Reverse code items}
\usage{
reverse.code(x, min.score, max.score, items)
}
\arguments{
\item{x}{Data frame containing items, some of which may need to be reverse coded}

\item{min.score}{Minimum possible score on the items}

\item{max.score}{Maximum possible score on the items}

\item{items}{A vector of item names or column numbers of items that need to be reverse coded}
}
\value{
A data frame
}
\description{
Reflects the scoring of items
}
\examples{
library(psychTools)
data(bfi)

## Perform principal components analysis to observe the signs of the component loadings
psych::pca(bfi[1:25], 5)
items <- c("A1", "C4", "C5", "E1", "E2", "O2", "O5")
bfi.new <- reverse.code(bfi[1:25], min.score = 1, max.score = 6, items)
psych::pca(bfi.new[1:25], 5)
}
