% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-itemselectr-spare.Rmd.
\name{Simple.Rasch}
\alias{Simple.Rasch}
\title{Rasch analysis of dichotomous or polytomous items}
\usage{
Simple.Rasch(raschdata, irtmodel)
}
\arguments{
\item{raschdata}{A data frame containing items where the lowest response category is scored as 0}

\item{irtmodel}{The irt model. "PCM", "PCM2" or "RSM"}
}
\value{
The output is a list containing the results of the analysis.
}
\description{
Basic Rasch analysis of dichotomous or polytomous items. This function is a wrapper for the tam.mml() function of the TAM package. The purpose of the function is to simplify the Rasch analysis process and to produce the output that users most typically want.
}
\examples{
library(TAM)
library(lordif)
data(Anxiety)
Anxiety <- Anxiety - 1 

## Fit the rating scale model
myoutput <- Simple.Rasch(raschdata = Anxiety[4:32], irtmodel = "RSM")
ls(myoutput)
myoutput$item.fit

#' ## Fit the partial credit model
myoutput <- Simple.Rasch(raschdata = Anxiety[4:32], irtmodel = "PCM2")
ls(myoutput)
myoutput$item.fit

}