% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANCOVA.dif.R
\name{ANCOVA.dif}
\alias{ANCOVA.dif}
\title{Performs a differential item functioning analysis of polytomous items using ANCOVA}
\usage{
ANCOVA.dif(x, Group, Trait)
}
\arguments{
\item{x}{A data frame containing the items. All the items should be scored in the same direction.}

\item{Group}{A factor indicating group membership}

\item{Trait}{A vector containing measures of the trait. This may be the total scores across all the items, the mean item scores, factor scores, or IRT person measures.}
}
\description{
Differential item functioning analysis of polytomous items via ANCOVA
}
\examples{
library(lordif)
data(Anxiety)

#'### Calculate the Trait score
Trait <- rowMeans(Anxiety[4:32])

#'### Perform the DIF analysis
out <- ANCOVA.dif(Anxiety[4:32], Anxiety$gender, Trait)
out
}
