% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANCOVA.dif.R, R/ANCOVA.dif2.R
\name{ANCOVA.dif}
\alias{ANCOVA.dif}
\title{Performs a differential item functioning analysis of polytomous items using ANCOVA}
\usage{
ANCOVA.dif(x, Group, Trait)

ANCOVA.dif(x, Group, Trait)
}
\arguments{
\item{x}{A data frame containing the items. All the items should be scored in the same direction.}

\item{Group}{A factor indicating group membership}

\item{Trait}{A vector containing measures of the trait. This may be the total scores across all the items, the mean item scores, factor scores, or IRT person measures.}
}
\description{
Differential item functioning analysis of polytomous items via ANCOVA

Differential item functioning analysis of polytomous items via ANCOVA
}
\examples{
library(lordif)
data(Anxiety)
x <- Anxiety

### Remove unwanted columns
x <- Anxiety[-c(1,3)]

### Calculate the Trait score
Trait <- rowMeans(Anxiety[2:30])

### Place the trait score in the first column, followed by the group, and then the items
x <- cbind(Trait, x)

### Perform the dif analysis
out <- ANCOVA.dif(x, Group, Trait) 
out
library(lordif)
data(Anxiety)
x <- Anxiety

### Remove unwanted columns
x <- Anxiety[-c(1,3)]

### Calculate the total score with simple summation
Trait <- rowMeans(Anxiety[2:30])

### Place the trait score in the first column, followed by the group, and then the items
x <- cbind(Trait, x)

### Perform the dif analysis
out <- ANCOVA.dif(x, Group, Trait)
out
}
