% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-itemselectr-spare.Rmd.
\name{dif.ancova}
\alias{dif.ancova}
\title{Performs a differential item functioning analysis of polytomous items using ANCOVA}
\usage{
dif.ancova(data, Trait, Group)
}
\arguments{
\item{data}{A data frame containing the items. All the items should be scored in the same direction.}

\item{Trait}{A vector containing measures of the trait. This may be the total scores across all the items, the mean item scores, factor scores, or IRT person measures.}

\item{Group}{A factor indicating group membership}
}
\description{
Differential item functioning analysis of polytomous items via ANCOVA
}
\examples{
library(lordif)
data(Anxiety)
### Perform the dif analysis
out <- dif.ancova(Anxiety[4:32], rowSums(Anxiety[4:32]), Anxiety$gender) 
out
}