% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.Rasch.R
\name{dif.Rasch}
\alias{dif.Rasch}
\title{Detection of statistically significant differential item functioning via ANCOVA of standardized Rasch residuals}
\usage{
dif.Rasch(std.residuals, Theta, Group)
}
\arguments{
\item{std.residuals}{A matrix or data frame containing the standardized residuals of a Rasch analysis}

\item{Theta}{A vector containing thetas (person measures)}

\item{Group}{A vector (factor) containing group membership}
}
\value{
DIF results
}
\description{
Detection of statistically significant differential item functioning via ANCOVA of standardized Rasch residuals
}
\examples{
library(TAM)
library(lordif)
Anxiety <- Anxiety - 1

## Fit the partial credit model
myRasch <- Simple.Rasch(Anxiety[ , 4:32], "PCM2")

## Perform the DIF analysis using the standardized residuals, Rasch person measure and group as input
dif.Rasch(myRasch$std.residuals, myRasch$thetas[,1], Anxiety$gender)
}
