% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-itemselectr-spare.Rmd.
\name{Rasch.DIF}
\alias{Rasch.DIF}
\title{Performs a differential item functioning analysis of polytomous items using ANCOVA}
\usage{
Rasch.DIF(std.res, group, thetas)
}
\arguments{
\item{std.res}{A data frame containing the standardized residuals of a Rasch analysis}

\item{group}{A factor indicating group membership}

\item{thetas}{A vector containing measures of the trait. This may be the total scores across all the items, the mean item scores, factor scores, or IRT person measures.}
}
\description{
Differential item functioning analysis of polytomous items via ANCOVA
}
\examples{
library(lordif)
data(Anxiety)
dd <- Simple.Rasch(Anxiety[4:32] - 1, "PCM2")
dd$std.residuals <- data.frame(dd$std.residuals)
### Perform the dif analysis
out <- Rasch.DIF(dd$std.residuals, Anxiety$gender, dd$thetas[,1]) 
out
}