% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-itemselectr-spare.Rmd.
\name{RaschItemAnalysis}
\alias{RaschItemAnalysis}
\title{Rasch analysis of polytomous items with the partial credit model}
\usage{
RaschItemAnalysis(x)
}
\arguments{
\item{x}{A data frame containing items where the lowest response category is scored as 0}
}
\value{
The output is a list containing the results of the analysis.
}
\description{
Basic Rasch analysis of polytomous items using the partial credit model. This function is a wrapper for the PCM() function of the eRm package. The purpose of the function is to simplify the Rasch analysis process and to produce the output that users most typically want.
}
\examples{
library(lordif)
data(Anxiety)
Anxiety <- Anxiety[4:32] - 1 

## Fit the partial credit model
myoutput <- RaschItemAnalysis(x = Anxiety)
ls(myoutput)
myoutput$Infit.outfit 
}