% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchepersCluster.R
\name{SchepersCluster}
\alias{SchepersCluster}
\title{Item selection based on inter-item covariances}
\usage{
SchepersCluster(x, nitems)
}
\arguments{
\item{x}{Data frame containing items that are all scored in the same direction}

\item{nitems}{The number of items to select (this value should be no lower than four)}
}
\value{
A list containing (a) the two items that constitute the core, (b) a data frame containing scale statistics, and (c) the names of the items that are included in the scale
}
\description{
A bottom-up item selection procedure based on inter-item covariances.
}
\examples{
library(lordif)
data(Anxiety)
SchepersCluster(Anxiety[4:32], 5)
}
